Day 14 + 20: Factor common cycle finding

+++ day-20/src/main.rs
@@ -2,13 +2,13 @@ use std::fs::File;
 use std::io::{Read, Seek};
 use day_20::parse::read_system;
 
-fn part1<R: Read + Seek>(input: R) -> Result<u64, aoc::Error> {
+fn part1<R: Read + Seek>(input: R) -> Result<usize, aoc::Error> {
     let mut system = read_system(input)?;
     let low_high = system.run_part1(1000);
     Ok(low_high.low * low_high.high)
 }
 
-fn part2<R: Read + Seek>(input: R) -> Result<u64, aoc::Error> {
+fn part2<R: Read + Seek>(input: R) -> Result<usize, aoc::Error> {
     let mut system = read_system(input)?;
     let answer = system.run_part2();
     Ok(answer)
